Fixed a series of cascading 32-bit pointer truncation crashes that and AINode_Stand on aarch64 Linux:
* BotGoalName: dropped (int) cast on returned levelitem name pointer
  — was sign-extending and segfaulting strlen inside sprintf in
  BotRecordNodeSwitch.
* BotRecordNodeSwitch: a2 retyped from int to const char *.  Original
  asm at .text 0x1001d3a0 reuses the function's own a2 param (saved on
  stack) as the 3rd %s arg to sprintf; IDA misread as 'v6 uninitialised
  local'.  All 14 callers had their (int)aSeekLtg etc. casts removed.
  intptr_t — they hold dword_10067504 + 28*node etc. (char* + int).
  vec3 pointers); two callers updated.
* AINode_Battle_Fight: replaced all stray 'a1' references (which
  resolved to the global 'char a1[2]="1";' weak symbol) with bs.
  IDA decompiled missing register-tracked locals as references to the
  global a1, leading to truncated bot_state pointer in BotChat_Kill,
  BotAttackMove, sub_100215E0 etc.
* BotAttackMove: a2 widened from int to intptr_t (bot_state ptr).
  WeaponJump: a2 (movestate ptr) widened to intptr_t; a3 also widened
* BotAimAtEnemy + BotCheckAttack: v3 widened from int to uintptr_t
  (holds sub_100354B0() weaponinfo_t* return; was truncated and then
* sub_1001CE20 extern in botlib_exports.c: first arg int -> intptr_t
  matching definition.

and chat.

Source-commit: ad5a046af2efac9e4abfb592e74c432c7886a381

+++ botlib/botlib_exports.c
@@ -35,6 +35,7 @@
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <errno.h>
 
 #include "../game/botlib.h"  /* bot_export_t, bot_import_t + prerequisite Q2 types */
@@ -96,7 +97,7 @@ extern int    BotClientSettings(int client, const void *settings);  /* 0x1002992
 extern int    BotSettings(int client, const void *settings);        /* 0x100299D0 */
 extern int    BotUpdateClient(int client, const void *buc);
 extern int    AAS_UpdateEntity(int entnum, float *state);
-extern int    sub_1001CE20(int origin, int ent, int channel, int soundindex,
+extern int    sub_1001CE20(intptr_t origin, int ent, int channel, int soundindex,
                            int volume_bits, int attn_bits, float timeofs);
 extern int    BotAddPointLight(_DWORD *origin, int ent, int radius_bits,
                                int r_bits, int g_bits, int b_bits,
@@ -454,7 +455,7 @@ int Export_BotAddSound(int *origin, int ent, int channel, int soundindex,
 {
     if (!BotLibSetup("BotUpdateSound")) return 1;
     if (!ValidEntityNumber(ent, "BotUpdateSound")) return 4;
-    return sub_1001CE20((int)origin, ent, channel, soundindex,
+    return sub_1001CE20((intptr_t)origin, ent, channel, soundindex,
                         *(int *)&volume,
                         *(int *)&attenuation,
                         timeofs);
